// src/model/Materia.tsx
export interface Materia {
    _ag?: boolean;
    _se: number;
    _cu: string;
    _di: string;
    _ap: number;
    _at: number;
    _pr: (number | string)[];
    _re: string;
    _el: boolean;
    _ho: boolean[][];
}
export interface Outro {
    _cu: string;
    _da: number[];
    _hd: string[][];
    _ap: number;

}